Improve identification of assertion methods
We were looking only at the name of the method. This means that if a custom assertion isn't correctly named it will miss an error. Similarly, we could get false positives where a method is named assert for some reason.

Now we check for an ExpectationFailedException. This is best because what defines an assertion method is that it throws that exception.

The problem now is that there are certain methods in PHPUnit that throw a base exception (e.g. the assertCount method) so when trying to use reflection on them, we will get the base exception.

These methods should be handled separately as they are odd cases.